software/antigravity/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
pydantic
pydantic-settings
pytest
pytest-asyncio
